requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.8.0

# OAuth
google-auth==2.26.2
google-auth-oauthlib==1.2.0
authlib==1.3.0

# Environment
python-dotenv==1.0.0

# Date & Time
python-dateutil==2.8.2

# Push Notifications
firebase-admin==6.4.0

# Utilities
aiofiles==23.2.1
httpx==0.26.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
ruff==0.1.13
